+++ src/lib/pages/proposal-details/components/ProposalTop.tsx
@@ -1,0 +1,5 @@
+import { Flex } from "@chakra-ui/react";
+
+export const ProposalTop = ({ id }: { id: number }) => {
+  return <Flex> {id} proposal top</Flex>;
+};