+++ CHANGELOG.md
@@ -47,6 +47,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- [#755](https://github.com/alleslabs/celatone-frontend/pull/755) Redirect from usei to home page
 - [#743](https://github.com/alleslabs/celatone-frontend/pull/743) Add mobile view for proposal list
 
 ### Bug fixes