+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- [#757](https://github.com/alleslabs/celatone-frontend/pull/757) api v1 - proposal data
 - [#731](https://github.com/alleslabs/celatone-frontend/pull/731) Add proposal detail page structure
 - [#749](https://github.com/alleslabs/celatone-frontend/pull/749) Add multi-type proposals
 - [#753](https://github.com/alleslabs/celatone-frontend/pull/753) api v1 - proposal type filter